ticket:1668 - bugfixes for polygon floodfill, query update for WKT in clip methods, restructuring of code in clip methods, and updating of test_clipping
Summary:
We have added support for clip(c, polygon((points)) ) and clip(c, subspace() ),
introduced many new tests in test_clipping, and improved the filling algorithm applied to polygons. Several design decisions were also made in the structure of the clipping methods, as well.

Test Plan: ./test.sh in test_clipping

Reviewers: dmisev, vmerticariu

Differential Revision: http://codereview.rasdaman.org/D532

+++ qlparser/qtunaryinduce.cc
@@ -248,7 +248,7 @@ QtUnaryInduce::computeUnaryMDDOp(QtMDD* operand, const BaseType* resultBaseType,
             resTile = NULL;
             delete mddres;
             mddres = NULL;
-            parseInfo.setErrorNo(static_cast<unsigned int>(err));
+            parseInfo.setErrorNo(err);
             throw parseInfo;
         }
 
@@ -322,7 +322,7 @@ QtUnaryInduce::computeUnaryOp(QtScalarData* operand, const BaseType* resultBaseT
             {
                 delete[] resultBuffer;
                 resultBuffer = NULL;
-                parseInfo.setErrorNo(static_cast<unsigned int>(err));
+                parseInfo.setErrorNo(err);
                 throw parseInfo;
             }
     }
